+++ gobblin-modules/gobblin-kafka-common/src/main/java/org/apache/gobblin/metrics/KafkaReportingFormats.java
@@ -31,8 +31,8 @@ import org.apache.gobblin.metrics.kafka.KafkaAvroEventReporter;
 import org.apache.gobblin.metrics.kafka.KafkaAvroReporter;
 import org.apache.gobblin.metrics.kafka.KafkaAvroSchemaRegistry;
 import org.apache.gobblin.metrics.kafka.KafkaEventReporter;
-import org.apache.gobblin.metrics.kafka.KeyValueEventObjectReporter;
-import org.apache.gobblin.metrics.kafka.KeyValueMetricObjectReporter;
+import org.apache.gobblin.metrics.reporter.KeyValueEventObjectReporter;
+import org.apache.gobblin.metrics.reporter.KeyValueMetricObjectReporter;
 import org.apache.gobblin.metrics.kafka.KafkaReporter;
 import org.apache.gobblin.metrics.kafka.PusherUtils;
 import org.apache.gobblin.metrics.reporter.util.KafkaAvroReporterUtil;
